requirements.txt
altair==5.0.1
folium==0.14.0
haversine==2.8.0
image==1.5.33
inflection==0.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.2.1
munkres==1.1.4
mypy-extensions==0.4.3
Pillow==9.4.0
pipreqs==0.4.13
pygame==2.4.0
plotly==5.15.0
pandas==1.5.3
scipy==1.10.0
streamlit==1.25.0
streamlit-folium==0.12.0